fix: removed the logic from middleware and added directly in page
